Cardano.CEM

Documentation

data AddressSpec

This is different ways to specify address

Constructors

ByAddress Address

ByPubKey PubKeyHash

BySameScript

addressSpecToAddress :: Address -> AddressSpec -> Address

data TxFanFilter script

Constructors

MkTxFanFilter

Fields

address :: AddressSpec

rest :: TxFanFilter' script

data TxFanFilter' script

Constructors

Anything

UnsafeBySameCEM BuiltinData
To be used via bySameCem

ByDatum BuiltinData

bySameCEM :: (ToData (State script), CEMScript script) => State script -> TxFanFilter' script

Constraint enforcing state of script mentioning this constraint

data Quantor

Constructors

Exist Integer

SumValueEq Value

data TxFanKind

Constructors

In

InRef

Out

data TxFanConstraint script

Constructors

MkTxFanC

Fields

txFanCKind :: TxFanKind

txFanCFilter :: TxFanFilter script

txFanCQuantor :: Quantor

class (HasSpine (Transition script), HasSpine (State script), Stages (Stage script), Show (Stage script)) => CEMScript script where

Associated Types

type Params script = params | params -> script

Params is immutable part of script Datum, | it should be used to encode all

type Stage script

Stage is datatype encoding all Intervals specified by script. | Stage logic is encoded by separate Stages type class. | It have separate StageParams datatype, | which is stored immutable in script Datum as well.

type State script = params | params -> script

State is changing part of script Datum. | It is in

type Transition script = transition | transition -> script

Transitions for deterministic CEM-machine

Methods

transitionStage :: Proxy script -> Map (Spine (Transition script)) (Stage script, Maybe (Spine (State script)), Maybe (Spine (State script)))

Each kind of Transition has statically associated Stage from/to States spines

transitionSpec :: Params script -> Maybe (State script) -> Transition script -> Either BuiltinString (TransitionSpec script)

data TransitionSpec script

Constructors

MkTransitionSpec

Fields

constraints :: [TxFanConstraint script]

signers :: [PubKeyHash]

getAllSpecSigners :: TransitionSpec script -> [PubKeyHash]

List of all signing keys required for transition spec

data CEMParams script

Static part of CEMScript datum. Datatype is actually used only by off-chain code due to Plutus limitations.

Constructors

MkCEMParams

Fields

scriptParams :: Params script

stagesParams :: StageParams (Stage script)

type CEMScriptDatum script = (StageParams (Stage script), Params script, State script)