Cardano.CEM.Monads

Documentation

data BlockchainParams

Params of blockchain required for transaction-building

Constructors

MkBlockchainParams

Fields

protocolParameters :: PParams LedgerEra

systemStart :: SystemStart

eraHistory :: LedgerEpochInfo

stakePools :: Set PoolId

class MonadFail m => MonadBlockchainParams m where

This monad gives access to all information about Cardano params, | which is various kind of Ledger params and ValidityBound/Slots semantics

Methods

askNetworkId :: m NetworkId

queryCurrentSlot :: m SlotNo

queryBlockchainParams :: m BlockchainParams

data UtxoQuery

Constructors

ByAddresses [Address]

ByTxIns [TxIn]

class MonadBlockchainParams m => MonadQueryUtxo m where

Ability to query current Utxo state of chain

Methods

queryUtxo :: UtxoQuery -> m (UTxO Era)

data ResolvedTx

Constructors

MkResolvedTx

Fields

txIns :: [(TxIn, TxInWitness)]

txInsReference :: [TxIn]

txOuts :: [TxOut CtxTx Era]

toMint :: TxMintValue BuildTx Era

interval :: Interval POSIXTime

additionalSigners :: [PubKeyHash]

signer :: ~(SigningKey PaymentKey)

data WrongSlotKind

Constructors

Early

Late

data TxSubmittingError

Constructors

WrongSlot WrongSlotKind Integer

TxInOutdated [TxIn]

UnhandledAutobalanceError (TxBodyErrorAutoBalance Era)

UnhandledNodeSubmissionError (ApplyTxError LedgerEra)

class MonadQueryUtxo m => MonadSubmitTx m where

Ability to send transaction to chain

Methods

submitResolvedTx :: ResolvedTx -> m (Either TxSubmittingError TxId)

class MonadSubmitTx m => MonadTest m where

Stuff needed to use monad for local testing

Methods

getTestWalletSks :: m [SigningKey PaymentKey]

List of keys having some amount of genesis ADA