Cardano.Extras

Description

Various utils to cope with `cardano-api` types Mainly stolen from `hydra-cardano-api` and some from atlas

Documentation

type Era = BabbageEra

type LedgerEra = BabbageEra StandardCrypto

type PlutusLang = PlutusScriptV2

plutusLang :: PlutusScriptVersion PlutusLang

plutusLangInEra :: ScriptLanguageInEra PlutusLang Era

unsafeHashFromBytes :: HashAlgorithm hash => ByteString -> Hash hash a

Parsing

Interpret some raw ByteString as a particular Hash.

NOTE: This throws if byte string has a length different that the expected target digest length.

parseSigningKeyTE :: ByteString -> Maybe (SigningKey PaymentKey)

toPlutusKeyHash :: Hash PaymentKey -> PubKeyHash

Conversions

signingKeyToPKH :: SigningKey PaymentKey -> PubKeyHash

signingKeyToAddress :: SigningKey PaymentKey -> Address

fromPlutusAddress :: NetworkId -> Address -> AddressInEra Era

addressInEraToAny :: AddressInEra Era -> AddressAny

unsafeScriptHashFromBytes :: ByteString -> ScriptHash

Unsafe wrap some bytes as a ScriptHash, relying on the fact that Plutus is using Blake2b_224 for hashing data (according to 'cardano-ledger').

Pre-condition: the input bytestring MUST be of length 28.

fromPlutusCurrencySymbol :: CurrencySymbol -> PolicyId

Convert a plutus CurrencySymbol into a cardano-api PolicyId.

fromPlutusValue :: Value -> Value

Convert a plutus Value into a cardano-api Value.

type TxInWitness = BuildTxWith BuildTx (Witness WitCtxTxIn Era)

Tx and other stuff construction

withKeyWitness :: TxIn -> (TxIn, TxInWitness)

Attaching mark meaning "TxIn would be witnessed by signing key"

mkInlineDatum :: ToData datum => datum -> TxOutDatum ctx Era

mkScriptWitness :: forall ctx. PlutusScript PlutusLang -> ScriptDatum ctx -> ScriptRedeemer -> ScriptWitness ctx Era

Construct a full script witness from a datum, a redeemer and a full PlutusScript. That witness has no execution budget.

toScriptData :: ToData a => a -> HashableScriptData

tokenToAsset :: TokenName -> AssetName

mkInlinedDatumScriptWitness :: ToData a => PlutusScript PlutusLang -> a -> BuildTxWith BuildTx (Witness WitCtxTxIn Era)

mintedTokens :: ToData redeemer => PlutusScript PlutusLang -> redeemer -> [(AssetName, Quantity)] -> TxMintValue BuildTx Era

txOutValue :: TxOut ctx Era -> Value

Fields

mTxOutDatum :: TxOut ctx Era -> Maybe HashableScriptData

utxoValue :: UTxO Era -> Value

cardanoModeParams :: ConsensusModeParams

Constants

Orphan instances

Show LedgerEpochInfo
Orphan Instances

Instance details

Methods

showsPrec :: Int -> LedgerEpochInfo -> ShowS

show :: LedgerEpochInfo -> String

showList :: [LedgerEpochInfo] -> ShowS

Monoid (TxOutValue Era)

Instance details

Methods

mempty :: TxOutValue Era

mappend :: TxOutValue Era -> TxOutValue Era -> TxOutValue Era

mconcat :: [TxOutValue Era] -> TxOutValue Era

Semigroup (TxOutValue Era)

Instance details

Methods

(<>) :: TxOutValue Era -> TxOutValue Era -> TxOutValue Era

sconcat :: NonEmpty (TxOutValue Era) -> TxOutValue Era

stimes :: Integral b => b -> TxOutValue Era -> TxOutValue Era

Eq (SigningKey PaymentKey)

Instance details

Methods

(==) :: SigningKey PaymentKey -> SigningKey PaymentKey -> Bool

(/=) :: SigningKey PaymentKey -> SigningKey PaymentKey -> Bool