Data.Spine

Documentation

class (Ord (Spine sop), Show (Spine sop)) => HasSpine sop where

Definitions

Spine is datatype, which tags constructors of ADT. | TH deriving utility generates Spines, which are Enums, | but one could introduce more complex Spine datatypes manually.

Associated Types

type Spine sop

Methods

getSpine :: sop -> Spine sop

deriveSpine :: Name -> Q [Dec]

Derives HasSpine | Usage: `$(deriveSpine ''HydraEvent)`

newtype OfSpine (x :: Spine datatype)

Newtype encoding sop value of fixed known spine

Constructors

UnsafeMkOfSpine

Fields

getValue :: datatype