Plutus.Deriving

Description

Taken from here and updated to new TH version: https://github.com/Liqwid-Labs/plutus-extra/blob/master/ plutus-derivingsrcPlutusTx/Deriving.hs

Documentation

deriveEq :: Name -> Q [Dec]

Generates a lawful Eq instance for the type named by the input. This instance will obey the following laws:

Reflexivity (for any x, x == x = True)

Symmetry (for any x, y, x == y = y PTx.== x)

Transitivity (for any x, y, z, if x == y and y == z, then x == z)

Substitution (for any x, y and pure f, x == y implies f x == f y)

Since: 1.0