Plutus.Extras

Documentation

type ValidatorType = BuiltinData -> BuiltinData -> BuiltinData -> ()

Signature of an untyped validator script.

wrapValidator :: (UnsafeFromData datum, UnsafeFromData redeemer, UnsafeFromData context) => (datum -> redeemer -> context -> Bool) -> ValidatorType

Wrap a typed validator to get the basic ValidatorType signature which can be passed to compile. REVIEW: There might be better ways to name this than "wrap"

scriptValidatorHash :: SerialisedScript -> ScriptHash

Compute the on-chain ScriptHash for a given serialised plutus script. Use this to refer to another validator script.

scriptCurrencySymbol :: SerialisedScript -> CurrencySymbol